+++ Modules/Statistics/LabelStatistics/vtkMRMLLabelStatisticsNode.h
@@ -31,9 +31,9 @@ class VTK_LABELSTATISTICS_EXPORT vtkMRMLLabelStatisticsNode : public vtkMRMLNode
   typedef struct LabelStatsEntry{
     int Label;
     int Count;
-    int Volume;
-    int Min;
-    int Max;
+    double Volume;
+    double Min;
+    double Max;
     double Mean;
     double StdDev;
   } LabelStatsEntry;
